Delay before pattern, instead of after
Problems were being encountered with trying to input the solution
immediately after the pattern is displayed. A large amount of the time,
the button would be pressed during the final half second sleep time, and
would not be registered. This change is a cheap fix to combat that
issue. It also has the advantage of offering a delay before the pattern
is shown, which feels less jarring than previously.